Added feature flag for d3 charts in skin.conf. Support for enableCrosshair.
