+++ skins/weewx-wdc/skin.conf
@@ -86,10 +86,12 @@ SKIN_VERSION = 2.3.3
 
     # What to show for the stat tiles.
     stat_tile_observations = outTemp, outHumidity, barometer, windSpeed, windDir, windGust, windGustDir, windrun, rain, rainRate, snowDepth, dewpoint, windchill, heatindex, UV, ET, radiation, appTemp, cloudbase, extraTemp1, extraHumid1, extraTemp2, extraHumid2, extraTemp3, extraHumid3, extraTemp4, extraHumid4, extraTemp5, extraHumid5, extraTemp6, extraHumid6, extraTemp7, extraHumid7, extraTemp8, extraHumid8
+
     # Stat tiles: Min/Max/Sum settings.
     stat_tiles_show_min = outTemp, outHumidity, barometer, pressure, altimeter, snowDepth, heatindex, dewpoint, windchill, cloudbase, appTemp
     stat_tiles_show_max = rainRate, hailRate, snowRate, UV
     stat_tiles_show_sum = rain, ET, hail, snow, lightning_strike_count, windrun
+
     # Stat tiles: show time when the min/max was reached.
     show_min_max_time_day = False
     show_min_max_time_yesterday = False
@@ -103,6 +105,8 @@ SKIN_VERSION = 2.3.3
     # Windrose.
     windRose_colors = "#f3cec9", "#e7a4b6", "#cd7eaf", "#a262a9", "#6f4d96", "#3d3b72"
     windRose_show_beaufort = True
+
+    ENABLE_D3_DIAGRAMS = True
 
     [[Icons]]
         #rain = "includes/icons/barometer.svg"
@@ -130,14 +134,18 @@ SKIN_VERSION = 2.3.3
 
         [[[combined_observations]]]
             [[[[climatogram_1]]]]
-                label = "Climatogram 1"
-                enableArea= True
+                label = "Gewummer 1"
                 curve="natural"
                 [[[[[obs]]]]]
                     [[[[[[outTemp]]]]]]
                         observation = "outTemp"
-                    [[[[[[rain]]]]]]
-                        observation = "rain"
+                    [[[[[[dew]]]]]]
+                        observation = "dewpoint"
+                        color = "#5F9EA0"
+                    [[[[[[barometer]]]]]]
+                        observation = "barometer"
+                    [[[[[[pressure]]]]]]
+                        observation = "pressure"
             [[[[climatogram_2]]]]
                 label = "Climatogram 2"
                 [[[[[obs]]]]]
@@ -224,6 +232,7 @@ SKIN_VERSION = 2.3.3
         [[[bar]]]
             enableLabel = False
             isInteractive = True
+            enableCrosshair = True
             yScaleOffset = 3
 
         # Observation specific settings.
@@ -356,6 +365,7 @@ SKIN_VERSION = 2.3.3
                 [[[[[tempdew]]]]]
                 [[[[[outHumidity]]]]]
                 [[[[[barometer]]]]]
+                [[[[[pressure]]]]]# @todo Remove
                 [[[[[windchill_heatindex]]]]]
                 [[[[[wind]]]]]
                 [[[[[windDir]]]]]
